vpn_prod/web-module/requirements.txt
aiosqlite>=0.19.0
aiohttp>=3.9.1
loguru>=0.7.2
flask[async]>=3.0.0
python-dotenv>=1.0.0
flask-login>=0.6.3
